unified module definition
