+++ Player/Decoder.js
@@ -1,4 +1,53 @@
-function d(a){throw a;}var g=void 0,i=!0,k=null,m=!1;function n(){return function(){}}var p;p||(p=eval("(function() { try { return Module || {} } catch(e) { return {} } })()"));var aa={},r;for(r in p)p.hasOwnProperty(r)&&(aa[r]=p[r]);var t="object"===typeof process&&"function"===typeof require,ba="object"===typeof window,ca="function"===typeof importScripts,da=!ba&&!t&&!ca;
+// universal module definition
+(function (root, factory) {
+    if (typeof define === 'function' && define.amd) {
+        // AMD. Register as an anonymous module.
+        define([], factory);
+    } else if (typeof exports === 'object') {
+        // Node. Does not work with strict CommonJS, but
+        // only CommonJS-like environments that support module.exports,
+        // like Node.
+        module.exports = factory();
+    } else {
+        // Browser globals (root is window)
+        root.Decoder = factory();
+    }
+}(this, function () {
+  
+
+  function error(message) {
+    console.error(message);
+    console.trace();
+  };
+
+  
+  function assert(condition, message) {
+    if (!condition) {
+      error(message);
+    };
+  };
+
+  
+  var getModule = function(_broadwayOnHeadersDecoded, _broadwayOnPictureDecoded){
+    
+    var windowBak;
+    if (typeof window != 'undefined'){
+      windowBak = window;
+    };
+    var window = {
+      _broadwayOnHeadersDecoded: _broadwayOnHeadersDecoded,
+      _broadwayOnPictureDecoded: _broadwayOnPictureDecoded
+    };
+    
+    /*
+    
+      The reason why this is all packed into one file is that this file can also function as worker.
+      you can integrate the file into your build system and provide the original file to be loaded into a worker.
+    
+    */
+    
+    
+    function d(a){throw a;}var g=void 0,i=!0,k=null,m=!1;function n(){return function(){}}var p;p||(p=eval("(function() { try { return Module || {} } catch(e) { return {} } })()"));var aa={},r;for(r in p)p.hasOwnProperty(r)&&(aa[r]=p[r]);var t="object"===typeof process&&"function"===typeof require,ba="object"===typeof window,ca="function"===typeof importScripts,da=!ba&&!t&&!ca;
 if(t){p.print||(p.print=function(a){process.stdout.write(a+"\n")});p.printErr||(p.printErr=function(a){process.stderr.write(a+"\n")});var fa=require("fs"),ga=require("path");p.read=function(a,b){var a=ga.normalize(a),c=fa.readFileSync(a);!c&&a!=ga.resolve(a)&&(a=path.join(__dirname,"..","src",a),c=fa.readFileSync(a));c&&!b&&(c=c.toString());return c};p.readBinary=function(a){return p.read(a,i)};p.load=function(a){ha(read(a))};p.thisProgram=1<process.argv.length?process.argv[1].replace(/\\/g,"/"):
 "unknown-program";p.arguments=process.argv.slice(2);"undefined"!==typeof module&&(module.exports=p);process.on("uncaughtException",function(a){a instanceof ia||d(a)})}else da?(p.print||(p.print=print),"undefined"!=typeof printErr&&(p.printErr=printErr),p.read="undefined"!=typeof read?read:function(){d("no read() available (jsc?)")},p.readBinary=function(a){if("function"===typeof readbuffer)return new Uint8Array(readbuffer(a));a=read(a,"binary");w("object"===typeof a);return a},"undefined"!=typeof scriptArgs?
 p.arguments=scriptArgs:"undefined"!=typeof arguments&&(p.arguments=arguments),this.Module=p,eval("if (typeof gc === 'function' && gc.toString().indexOf('[native code]') > 0) var gc = undefined")):ba||ca?(p.read=function(a){var b=new XMLHttpRequest;b.open("GET",a,m);b.send(k);return b.responseText},"undefined"!=typeof arguments&&(p.arguments=arguments),"undefined"!==typeof console?(p.print||(p.print=function(a){console.log(a)}),p.printErr||(p.printErr=function(a){console.log(a)})):p.print||(p.print=
@@ -186,4 +235,339 @@ typeof j&&j.stack)&&p.fa("exception thrown: "+[j,j.stack]),d(j)))}finally{}};
 function mc(a){function b(){if(!p.calledRun&&(p.calledRun=i,!H)){Sa||(Sa=i,Na(R));Na(Pa);ba&&jc!==k&&p.fa("pre-main prep time: "+(Date.now()-jc)+" ms");if(p.onRuntimeInitialized)p.onRuntimeInitialized();p._main&&lc&&p.callMain(a);if(p.postRun)for("function"==typeof p.postRun&&(p.postRun=[p.postRun]);p.postRun.length;)Ua(p.postRun.shift());Na(Ra)}}a=a||p.arguments;jc===k&&(jc=Date.now());if(!(0<S)){if(p.preRun)for("function"==typeof p.preRun&&(p.preRun=[p.preRun]);p.preRun.length;)Ta(p.preRun.shift());
 Na(Oa);!(0<S)&&!p.calledRun&&(p.setStatus?(p.setStatus("Running..."),setTimeout(function(){setTimeout(function(){p.setStatus("")},1);b()},1)):b())}}p.run=p.Ng=mc;function nc(a){p.noExitRuntime||(H=i,y=ic,Na(Qa),t?(process.stdout.once("drain",function(){process.exit(a)}),console.log(" "),setTimeout(function(){process.exit(a)},500)):da&&"function"===typeof quit&&quit(a),d(new ia(a)))}p.exit=p.hg=nc;
 function A(a){a&&(p.print(a),p.fa(a));H=i;d("abort() at "+Fa()+"\nIf this abort() is unexpected, build with -s ASSERTIONS=1 which can give more information.")}p.abort=p.abort=A;if(p.preInit)for("function"==typeof p.preInit&&(p.preInit=[p.preInit]);0<p.preInit.length;)p.preInit.pop()();var lc=m;p.noInitialRun&&(lc=m);mc();
+
+
+    
+    
+    
+    
+    
+    
+    var resultModule = window.Module || this.Module;
+    return resultModule;
+  };
+  
+  var Broadway = function(parOptions){
+    this.options = parOptions;
+    
+    var Module = getModule(function () {
+
+    }, function ($buffer, width, height) {
+      var buffer = this.pictureBuffers[$buffer];
+      if (!buffer) {
+        buffer = this.pictureBuffers[$buffer] = toU8Array($buffer, (width * height * 3) / 2);
+      }
+      this.onPictureDecoded(buffer, width, height);
+    }.bind(this));
+
+    var HEAP8 = Module.HEAP8;
+    var HEAPU8 = Module.HEAPU8;
+    var HEAP16 = Module.HEAP16;
+    var HEAP32 = Module.HEAP32;
+    var _h264bsdClip = Module._get_h264bsdClip();
+
+    
+    var MAX_STREAM_BUFFER_LENGTH = 1024 * 1024;
+  
+    // from old constructor
+    Module._broadwayInit();
+    
+    /**
+   * Creates a typed array from a HEAP8 pointer. 
+   */
+    function toU8Array(ptr, length) {
+      return HEAPU8.subarray(ptr, ptr + length);
+    };
+    this.streamBuffer = toU8Array(Module._broadwayCreateStream(MAX_STREAM_BUFFER_LENGTH), MAX_STREAM_BUFFER_LENGTH);
+    this.pictureBuffers = {};
+    
+    this.onPictureDecoded = function (buffer, width, height) {
+      
+    };
+    
+    /**
+     * Decodes a stream buffer. This may be one single (unframed) NAL unit without the
+     * start code, or a sequence of NAL units with framing start code prefixes. This
+     * function overwrites stream buffer allocated by the codec with the supplied buffer.
+     */
+    this.decode = function decode(buffer) {
+      // console.info("Decoding: " + buffer.length);
+      this.streamBuffer.set(buffer);
+      Module._broadwaySetStreamLength(buffer.length);
+      Module._broadwayPlayStream();
+    };
+
+
+    
+    function patchOptimizations(config, patches) { 
+      var scope = getGlobalScope();
+      for (var name in patches) {
+        var patch = patches[name];
+        if (patch) {
+          var option = config[name];
+          if (!option) option = "original";
+          console.info(name + ": " + option);
+          assert (option in patch.options);
+          var fn = patch.options[option].fn;
+          if (fn) {
+            scope[patch.original] = Module.patch(null, patch.name, fn);
+            console.info("Patching: " + patch.name + ", with: " + option);
+          }
+        }
+      }
+    };
+    
+    var patches = {
+      "filter": {
+        name: "_h264bsdFilterPicture",
+        display: "Filter Picture",
+        original: "Original_h264bsdFilterPicture",
+        options: {
+          none: {display: "None", fn: function () {}},
+          original: {display: "Original", fn: null},
+        }
+      },
+      "filterHorLuma": {
+        name: "_FilterHorLuma",
+        display: "Filter Hor Luma",
+        original: "OriginalFilterHorLuma",
+        options: {
+          none: {display: "None", fn: function () {}},
+          original: {display: "Original", fn: null},
+          optimized: {display: "Optimized", fn: OptimizedFilterHorLuma}
+        }
+      },
+      "filterVerLumaEdge": {
+        name: "_FilterVerLumaEdge",
+        display: "Filter Ver Luma Edge",
+        original: "OriginalFilterVerLumaEdge",
+        options: {
+          none: {display: "None", fn: function () {}},
+          original: {display: "Original", fn: null},
+          optimized: {display: "Optimized", fn: OptimizedFilterVerLumaEdge}
+        }
+      },
+      "getBoundaryStrengthsA": {
+        name: "_GetBoundaryStrengthsA",
+        display: "Get Boundary Strengths",
+        original: "OriginalGetBoundaryStrengthsA",
+        options: {
+          none: {display: "None", fn: function () {}},
+          original: {display: "Original", fn: null},
+          optimized: {display: "Optimized", fn: OptimizedGetBoundaryStrengthsA}
+        }
+      }
+    };
+    function getGlobalScope() {
+      return function () { return this; }.call(null);
+    };
+    
+    /* Optimizations */
+
+    function clip(x, y, z) {
+      return z < x ? x : (z > y ? y : z);
+    }
+
+    function OptimizedGetBoundaryStrengthsA($mb, $bS) {
+      var $totalCoeff = $mb + 28;
+
+      var tc0 = HEAP16[$totalCoeff + 0 >> 1];
+      var tc1 = HEAP16[$totalCoeff + 2 >> 1];
+      var tc2 = HEAP16[$totalCoeff + 4 >> 1];
+      var tc3 = HEAP16[$totalCoeff + 6 >> 1];
+      var tc4 = HEAP16[$totalCoeff + 8 >> 1];
+      var tc5 = HEAP16[$totalCoeff + 10 >> 1];
+      var tc6 = HEAP16[$totalCoeff + 12 >> 1];
+      var tc7 = HEAP16[$totalCoeff + 14 >> 1];
+      var tc8 = HEAP16[$totalCoeff + 16 >> 1];
+      var tc9 = HEAP16[$totalCoeff + 18 >> 1];
+      var tc10 = HEAP16[$totalCoeff + 20 >> 1];
+      var tc11 = HEAP16[$totalCoeff + 22 >> 1];
+      var tc12 = HEAP16[$totalCoeff + 24 >> 1];
+      var tc13 = HEAP16[$totalCoeff + 26 >> 1];
+      var tc14 = HEAP16[$totalCoeff + 28 >> 1];
+      var tc15 = HEAP16[$totalCoeff + 30 >> 1];
+
+      HEAP32[$bS + 32 >> 2] = tc2 || tc0 ? 2 : 0;
+      HEAP32[$bS + 40 >> 2] = tc3 || tc1 ? 2 : 0;
+      HEAP32[$bS + 48 >> 2] = tc6 || tc4 ? 2 : 0;
+      HEAP32[$bS + 56 >> 2] = tc7 || tc5 ? 2 : 0;
+      HEAP32[$bS + 64 >> 2] = tc8 || tc2 ? 2 : 0;
+      HEAP32[$bS + 72 >> 2] = tc9 || tc3 ? 2 : 0;
+      HEAP32[$bS + 80 >> 2] = tc12 || tc6 ? 2 : 0;
+      HEAP32[$bS + 88 >> 2] = tc13 || tc7 ? 2 : 0;
+      HEAP32[$bS + 96 >> 2] = tc10 || tc8 ? 2 : 0;
+      HEAP32[$bS + 104 >> 2] = tc11 || tc9 ? 2 : 0;
+      HEAP32[$bS + 112 >> 2] = tc14 || tc12 ? 2 : 0;
+      HEAP32[$bS + 120 >> 2] = tc15 || tc13 ? 2 : 0;
+
+      HEAP32[$bS + 12 >> 2] = tc1 || tc0 ? 2 : 0;
+      HEAP32[$bS + 20 >> 2] = tc4 || tc1 ? 2 : 0;
+      HEAP32[$bS + 28 >> 2] = tc5 || tc4 ? 2 : 0;
+      HEAP32[$bS + 44 >> 2] = tc3 || tc2 ? 2 : 0;
+      HEAP32[$bS + 52 >> 2] = tc6 || tc3 ? 2 : 0;
+      HEAP32[$bS + 60 >> 2] = tc7 || tc6 ? 2 : 0;
+      HEAP32[$bS + 76 >> 2] = tc9 || tc8 ? 2 : 0;
+      HEAP32[$bS + 84 >> 2] = tc12 || tc9 ? 2 : 0;
+      HEAP32[$bS + 92 >> 2] = tc13 || tc12 ? 2 : 0;
+      HEAP32[$bS + 108 >> 2] = tc11 || tc10 ? 2 : 0;
+      HEAP32[$bS + 116 >> 2] = tc14 || tc11 ? 2 : 0;
+      HEAP32[$bS + 124 >> 2] = tc15 || tc14 ? 2 : 0;
+    }
+
+    function OptimizedFilterVerLumaEdge ($data, bS, $thresholds, imageWidth) {
+      var delta, tc, tmp;
+      var p0, q0, p1, q1, p2, q2;
+      var tmpFlag;
+      var $clp = _h264bsdClip + 512;
+      var alpha = HEAP32[$thresholds + 4 >> 2];
+      var beta = HEAP32[$thresholds + 8 >> 2];
+      var val;
+
+      if (bS < 4) {
+        tmp = tc = HEAPU8[HEAP32[$thresholds >> 2] + (bS - 1)] & 255;
+        for (var i = 4; i > 0; i--) {
+          p1 = HEAPU8[$data + -2] & 255;
+          p0 = HEAPU8[$data + -1] & 255;
+          q0 = HEAPU8[$data] & 255;
+          q1 = HEAPU8[$data + 1] & 255;
+          if ((Math.abs(p0 - q0) < alpha) && (Math.abs(p1 - p0) < beta) && (Math.abs(q1 - q0) < beta)) {
+            p2 = HEAPU8[$data - 3] & 255;
+            if (Math.abs(p2 - p0) < beta) {
+              val = (p2 + ((p0 + q0 + 1) >> 1) - (p1 << 1)) >> 1;
+              HEAP8[$data - 2] = p1 + clip(-tc, tc, val);
+              tmp++;
+            }
+
+            q2 = HEAPU8[$data + 2] & 255;
+            if (Math.abs(q2 - q0) < beta) {
+              val = (q2 + ((p0 + q0 + 1) >> 1) - (q1 << 1)) >> 1;
+              HEAP8[$data + 1] = (q1 + clip(-tc, tc, val));
+              tmp++;
+            }
+
+            val = ((((q0 - p0) << 2) + (p1 - q1) + 4) >> 3);
+            delta = clip(-tmp, tmp, val);
+
+            p0 = HEAPU8[$clp + (p0 + delta)] & 255;
+            q0 = HEAPU8[$clp + (q0 - delta)] & 255;
+            tmp = tc;
+            HEAP8[$data - 1] = p0;
+            HEAP8[$data] = q0;
+
+            $data += imageWidth;
+          }
+        }
+      } else {
+        OriginalFilterVerLumaEdge($data, bS, $thresholds, imageWidth);
+      }
+    }
+
+    /**
+ * Filter all four successive horizontal 4-pixel luma edges. This can be done when bS is equal to all four edges.
+ */
+    function OptimizedFilterHorLuma ($data, bS, $thresholds, imageWidth) {
+      var delta, tc, tmp;
+      var p0, q0, p1, q1, p2, q2;
+      var tmpFlag;
+      var $clp = _h264bsdClip + 512;
+      var alpha = HEAP32[$thresholds + 4 >> 2];
+      var beta = HEAP32[$thresholds + 8 >> 2];
+      var val;
+
+      if (bS < 4) {
+        tmp = tc = HEAPU8[HEAP32[$thresholds >> 2] + (bS - 1)] & 255;
+        for (var i = 16; i > 0; i--) {
+          p1 = HEAPU8[$data + (-imageWidth << 1)] & 255;
+          p0 = HEAPU8[$data + -imageWidth] & 255;
+          q0 = HEAPU8[$data] & 255;
+          q1 = HEAPU8[$data + imageWidth] & 255;
+
+          if ((Math.abs(p0 - q0) < alpha) && (Math.abs(p1 - p0) < beta) && (Math.abs(q1 - q0) < beta)) {
+            p2 = HEAPU8[$data + (-imageWidth * 3)] & 255;
+            if (Math.abs(p2 - p0) < beta) {
+              val = (p2 + ((p0 + q0 + 1) >> 1) - (p1 << 1)) >> 1;
+              HEAP8[$data + (-imageWidth << 1)] = p1 + clip(-tc, tc, val);
+              tmp++;
+            }
+
+            q2 = HEAPU8[$data + (imageWidth << 2)] & 255;
+            if (Math.abs(q2 - q0) < beta) {
+              val = (q2 + ((p0 + q0 + 1) >> 1) - (q1 << 1)) >> 1;
+              HEAP8[$data + imageWidth] = (q1 + clip(-tc, tc, val));
+              tmp++;
+            }
+
+            val = ((((q0 - p0) << 2) + (p1 - q1) + 4) >> 3);
+            delta = clip(-tmp, tmp, val);
+
+            p0 = HEAPU8[$clp + (p0 + delta)] & 255;
+            q0 = HEAPU8[$clp + (q0 - delta)] & 255;
+            tmp = tc;
+            HEAP8[$data - imageWidth] = p0;
+            HEAP8[$data] = q0;
+
+            $data ++;
+          }
+        }
+      } else {
+        OriginalFilterHorLuma($data, bS, $thresholds, imageWidth);
+      }
+    }
+  };
+
+  
+  Broadway.prototype = {
+    configure: function (config) {
+      // patchOptimizations(config, patches);
+      console.info("Broadway Configured: " + JSON.stringify(config));
+    }
+    
+  };
+  
+  
+  if (typeof self != "undefined"){
+    var isWorker = false;
+    var decoder;
+    self.addEventListener('message', function(e) {
+      
+      if (isWorker){
+        decoder.decode(new Uint8Array(e.data));
+        
+      }else{
+        if (e.data === "Broadway.js - Worker init"){
+          isWorker = true;
+          decoder = new Broadway();
+          decoder.onPictureDecoded = function (buffer, width, height) {
+            if (buffer) {
+              buffer = new Uint8Array(buffer);
+            };
+
+            // post dimensions seperately
+            postMessage({width: width, height: height});
+
+            // buffer needs to be copied because we give up ownership
+            var copyU8 = new Uint8Array(buffer.length);
+            copyU8.set( buffer, 0, buffer.length );
+            
+            // only post the buffer (slightly faster)
+            // add 2nd parameter to indicate transfer of owner ship (this it was makes this worker implementation faster)
+            postMessage(copyU8.buffer, [copyU8.buffer]);
+
+          };
+          postMessage({consoleLog: "initialized" });
+        };
+      };
+
+
+    }, false);
+  };
+  
+  
+  // Just return a value to define the module export.
+  // This example returns an object, but the module
+  // can return a function as the exported value.
+  return Broadway;
+  
+}));
