+++ Player/treeDemo.html
@@ -3,15 +3,13 @@
   <link type="text/css" href="screen.css" rel="stylesheet" />
 </head>
 <body onload="load()">
-  <script src="sylvester.js" type="text/javascript"></script>
-  <script src="glUtils.js" type="text/javascript"></script>
+  <script type="text/javascript" src="Decoder.js"></script>
+  <script type="text/javascript" src="YUVWebGLCanvas.js"></script>
+  <script type="text/javascript" src="Player.js"></script>
   
-  <script type="text/javascript" src="util.js"></script>
   <script type="text/javascript" src="stream.js"></script>
-  <script type="text/javascript" src="avc-codec.js"></script>
-  <script type="text/javascript" src="avc.js"></script>
+  
   <script type="text/javascript" src="mp4.js"></script>
-  <script src='canvas.js' type='text/javascript'></script>
   
   <script type="text/javascript">
     function load() {
